# Floating minmax: fix negative zero handling and dedicated test coverage for arrays of +0.0 and -0.0 only