--- a/stl/inc/xutility
+++ b/stl/inc/xutility
@@ -6801,14 +6801,14 @@ constexpr bool _Is_min_max_optimization_safe = // Activate the vector algorithms
 #endif // _HAS_CXX20
             is_same<_Pr, less<>>, is_same<_Pr, less<_Elem>>>>; // predicate is less
 
-// The value-based vectorized rather than the position-based one does not always produce
-// the expected results for floatting point types.
+// Unlike the position-based vectorized implementation, the value-based vectorized implementation
+// does not always produce the expected results for floating-point types.
 //
-// Efficient vectorization needs to find vertical minmax first, and then the horizontal one.
-// This alters order of comparison: index zero element is first compared against
+// Efficient vectorization needs to find the vertical minmax first, and then the horizontal one.
+// This alters the order of comparison: index zero element is first compared against
 // vector size equal index element and only in the end against index one element.
 // With equivalent but distinguishable +0.0 and -0.0 values, the altered comparison order
-// will not produce the expected result in some cases (will return +0.0 instead of -0.0 or the reverse)
+// will not produce the expected result in some cases (will return +0.0 instead of -0.0 or the reverse).
 //
 // The result is still acceptable for /fp:fast when +0.0 / -0.0 are not expected to be properly distinguished,
 // and the compiler itself takes advantage of it.