--- a/tests/std/include/test_min_max_element_support.hpp
+++ b/tests/std/include/test_min_max_element_support.hpp
@@ -5,6 +5,7 @@
 
 #include <algorithm>
 #include <cassert>
+#include <cmath>
 #include <cstddef>
 #include <type_traits>
 #include <utility>