--- a/tests/std/include/test_vector_algorithms_support.hpp
+++ b/tests/std/include/test_vector_algorithms_support.hpp
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
+#pragma once
+
 #include <isa_availability.h>
 #include <random>
 #include <vector>