… tests/std/test.lst
… tests/std/tests/VSO_0000000_vector_algorithms/test.cpp
… tests/std/tests/VSO_0000000_vector_algorithms_floats/env.lst
… tests/std/tests/VSO_0000000_vector_algorithms_floats/test.cpp