@AlexGuteniev 2024-06-19
Fixing the behavior is blocked on the compiler bug DevCom-10686775