@StephanTLavavej 2024-06-20
:warning: @AlexGuteniev mentioned that there will be a stealth merge conflict with my #4741, where `vector_algorithms.cpp` is testing `_M_IX86_FP`.