@AlexGuteniev 2024-06-20
No significant performance changes.
I expect AVX codegen to be only slightly worse.

```
out\bench\benchmark-minmax_element.exe --benchmark_filter=(float^|double) --benchmark_min_time=1s
```

Benchmark                        |   before      |     after
---------------------------------|---------------|--------------
bm<float, 8021, Op::Min>         |    1182 ns    |     1197 ns
bm<float, 8021, Op::Max>         |    1215 ns    |     1193 ns
bm<float, 8021, Op::Both>        |    1376 ns    |     1376 ns
bm<float, 8021, Op::Min_val>     |     917 ns    |      893 ns
bm<float, 8021, Op::Max_val>     |     904 ns    |      893 ns
bm<float, 8021, Op::Both_val>    |     949 ns    |      957 ns
bm<double, 8021, Op::Min>        |    2219 ns    |     2409 ns
bm<double, 8021, Op::Max>        |    2373 ns    |     2472 ns
bm<double, 8021, Op::Both>       |    2774 ns    |     2769 ns
bm<double, 8021, Op::Min_val>    |    1888 ns    |     1920 ns
bm<double, 8021, Op::Max_val>    |    1868 ns    |     1878 ns
bm<double, 8021, Op::Both_val>   |    1932 ns    |     1903 ns