@StephanTLavavej 2024-08-25
I had to push an additional commit to drop my eternal nemeses:

```
C:\Temp>cl /EHsc /nologo /W4 /fp:strict meow.cpp
meow.cpp

C:\Temp>cl /clr /nologo /W4 /fp:strict meow.cpp
cl : Command line error D8016 : '/clr' and '/fp:strict' command-line options are incompatible

C:\Temp>cl /clr:pure /nologo /W4 /fp:strict meow.cpp
cl : Command line warning D9035 : option 'clr:pure' has been deprecated and will be removed in a future release
cl : Command line error D8016 : '/clr:pure' and '/fp:strict' command-line options are incompatible
```